BAU: Move CRI return lambdas to a step function
This moves the three lambdas associated with processing a users callback
from a CRI to a step function. This will prevent bouncing back to the
frontend.

It also includes the process journey step lambda, which is called
between each of the lambdas doing the processing. In a future iteration
we may be able to drop calling this lambda and just move from one lambda
to the next.

Other interesting things worth noting:
- There are two endpoints defined in the private openAPI spec which point
at the new step function. This is to allow a seamless refactor of the
frontend. Initially the frontend will invoke the step function at the
`/journey/cri/validate-callback` endpoint. This is the first endpoint it
currently calls for a users callback, and so is the natural entrypoint
for the step function. Once this change is deployed, we can update the
frontend to use the more appropriate `/journey/cri/callback` endpoint
and drop the old one.
- The requestTemplates in the openAPI spec contains two mapping
  templates. One for `application/json` requests and another for
`application/x-www-form-urlencoded` requests. Currently the frontend
sends `x-www-form-urlencoded`. Having both defined will allow us to
refactor the frontend to use `application/json` without breaking things.
- The openAPI template mapping for the `/journey/{journeyStep}` endpoint
  has been updated to pass the correct format JSON to the now refactored
ProcessJourneyStep function
- The `ValidateOauthCallbackHandler` now implements
  `RequestStreamHandler`, rather than having an input object. This is to
allow us to use Jackson annotations in the `CredentialIssuerRequestDto`
object. This is neccessary as there is mismatch in the format of the
property names we pass from the frontend and deserialize in the backend.
We can use `@JsonProperty` to smooth over this inconsistency. It would be
nicer to use the inbuilt deserialization of the lambda runtime and
specify the input as a `CredentialIssuerRequestDto` object.
Unfortunately the inbuilt deserializtion can't make use of annotations.
So we've fallen back to using input and output streams to manage the
deserialization ourselves. A future piece of work will be to update the
frontend to use the same property names as the backend expects, at which
point we can use the inbuilt deserialization again.