BAU: Flag to disable warning of null Dynamo item
Overload the `getItem()` method of the `DataStore` class to provide a flag that will indicate if returning null should log a warning. This defaults to logging the warning.

It is not common for us to retrieve an item from DynamoDb when we don't expect it to be there. We therefore log a warning when that happens.
However there is a scenario where dynamo returning null is expected. This is when we check if a jwt id has previously been used. In that case we do not want to log a warning, as it is expected behaviour.