BAU: Add permission boundary to step functions
The role the secure pipeline uses to create IAM roles can only create
them if they have a permission boundary attached.

The `JourneyEngineStepFunction` resource was created before we started
using the secure pipeline, so was deployed succesfully.

This adds the permissions boundary to that old step function as well as
the new one. The step functions get created with an execution role,
which is why we need this.