BAU: Directly deserialize input in validate-oauth-callback lambda
Previously the frontend was sending the oauth parameters to the backend
in `snake_case`. The java object we use to represent the request uses
field names with camel case. This mismatch meant that we needed to use
the `@JsonProperty` annotation to map one to the other.

When defining the input type for a lambda handler you can not rely on
Jackson annotations. If there is a mismatch between the Jackson version
the lambda runtime is using under the hood, and the version you're using
for your annotations, they won't be applied on deserialization. This
resulted in objects with all null values.

To get around this, the validate-oauth-callback lambda was receiving the
input as a stream and then handling the deserialisation within the
lambda.

The frontend has now been updated to send the parameters as a JSON body,
and matching the format expected in the backend. This means we can now
use the inbuilt deserialization of the lambda runtime.