BAU: Call next lambdas directly in CRI return step function
This updates the lambda in the CRI return step function to call each
other directly, rather than by calling the process journey step lambda.

This does mean that we'll lose a couple of states, namely
`RETRIEVE_CRI_OAUTH_ACCESS_TOKEN` and `RETRIEVE_CRI_CREDENTIAL`, but I'm
not sure that it matters much. If there are errors we need to
investigate we'll be able to see from the lambda logs where the journey
broke down.