PYIC-2075: Send audit event when GPG45 profile matched
We currently don't send an audit message when a user has matched against
a GPG45 profile.

This adds that audit event, with details of the profile, the scores
achieved, and the txn ID's from the VC's used.

We send the full evidence block of the VCs in audit events when we fetch
them from CRIs, so the txn's here can be used to find those audit
events and get more details.