BAU: Rename UserIssuedCredentialsItem
Rename `UserIssuedCredentialItem` to `VcStoreItem`. This is a mostly mechanical IDE dririven rename refactor.

The UserIssuedCredentialsItem is the DynamoDB item for the "Vc Store". We will be refactoring hwo we handle various VCs in the system and we want to make sure that we are dealing with the wrapping DynamoDB item here.