BAU: Move Debug Credentials handling out of UserIdentityService
Move all handling of "Debug Credentials" out of the UserIdentityService and up into the BuildDebugCredentials lambda.

The "debug credentials" are only used by the BuildDebugCredentials lambda called directly when constructing the debug page. We don't need to have this in the UserIdentityService code and tests. Ultimatly the Debug page and the lambda will be removed, so I have not done any farther refactoring here for now.