BAU: Retry getting session by token
When exchanging an authorisation code for an access token for our clients to access an completed identity we add the token to the session. This is then used as a global secondary index. We use that index to retrieve the session when the token is presented to us.

However we occasionally see that the token is not found. We think that this is because the token added to the session is not yet available in the secondary index.

Since this is a secondary index, we cannot query DynamoDB with a consistent read.

To prevent dropping the request when it will likely be available soon, we have implemented a basic backoff and retry when getting the session using the access token.

This is a small tactical fix. We should look at the overall architecture of token handling and session handling to see if a better strategic solution is available.