BAU: Handle fraud credentials containing activity
VC's received from the fraud CRI may start to include an activity score
as well as a fraud score. This change allows the GPG 45 evaluator to
handle both versions.

A credential that contains both fraud and activity scores is split into
two so the different scores can be compared with other scores of the
same type. We do a similar thing for DCMAW and f2f credentials.