BAU: Simplify GPG45 score comparators
The comparators were comparing scores based on the score and then by the
number of CIs a VC contained. CI checking is handled separately now so
this is no longer required. It also doesn't make sense where we have VCs
containing multiple scores. If a VC had a fraud score and an activity
history score, which score would the CI apply to?