BAU: Add f2f check validator
We have logic to check if VC's are deemed "successful". This is used
when deciding where to send a user, but also for deciding which VCs we
should extract data from when constructing an identity.

We need this for the new f2f VC as it will be the only VC that contains
the name and date of birth data.