BAU: Refactor process journey step state machine init
The state machine config used by the process journey step function was
being loaded from file on every invocation of the lambda. As well as
this, a new ObjectManager was being instantiated to do this. These are
potentially heavy operations which could slow down a lambda.

This moves the loading of the file to the initialization phase. This
might result in a slightly slower cold start however should be
beneficial in the long run. If we're using provisioned concurrency it's
also not an issue.

This does enforce the use of the `ipv-core-main-journey` journey type,
however this is currently the only journey type we have. If we introduce
more (potentially with mitigations work) we can look at how to include
more journey types. Possible loading them in the initialization flag and
then using them in the handler based on the journey type value in the
users session.