BAU: Apply singleton pattern to ConfigService
It might be useful for certain classes to have direct access to a
ConfigService without it having to be passed down through a chain of
method calls, and without having to instantiate a whole new instance of
it for themselves.

Applying the singleton pattern like this would allow some class in the
depths of the system to use an already existing config service. This
would hopefully help with performance of the lambda, and also benefit
from the parameter caching we're already doing.

There can be issues with this pattern when in a multithreaded
environment, but I don't think this is an issue when running in the
context of a lambda.