PYIC-3010: Load statemachines and select from session
This adds logic to the process-journey-step state machine to load
statemachine files for the journey types that we care about. This is now
the main journey that users are currently running, as well as the
refactor journey that we're building.

The lambda will then use the state machine for the correct journey type,
set in their session when it was initialized. If a matching state
machine can't be found (which shouldn't ever happen), we log it and
throw an error.