PYIC-3011: Move evaluate-gpg45-scores lambda to criReturn step function
This adds the evaluate-gpg45-scores lambda into the criReturn step
function, and runs it after the retrieve-cri-credential lambda IF the
user is on an 'ipv-core-refactor-journey' type journey.

This change allows us to refactor the journey state machine files
without affecting the current main user journey.

The retrieve-cri-credential lambda has had its output instruments with
the users journey type from their session. This is the used in the step
function to decide if it should run the evaluate-gpg45-scores lambda.
This instrumentation and logic will be removed once the refactor work
has been completed.

This change has been deployed to my dev account, and tested with each
type of user journey.