PYIC-3012: Create new CRI event type for state machine
This new event type does a couple of things.

First it adds a bit of sugar on creating the journey response for a CRI,
to keep the statemachine file a bit cleaner.

Second, it adds a new `checkIfDisabled` property which defines a map of
CRI IDs to check for their "enabledness", and transition to a different
state/event response depending on that "enabledness".

The map is a LinkedHashMap, so will preserve the ordering of the
entries. This means that the ordering of the CRIs to be checked in the
definition is important. The first disabled CRI that's found is used.

Having a map of CRIs to check (rather than just checking the CRI that
this event is defined for) allows for a transition to occur if a CRI
further along the current journey is disabled. Equally, this map may
just contain once entry which is for the CRI defined in the main part of
the event.