PYIC-3013: Add journeys to ipv-core-refactor-journey statemachine
We have four main journeys through the system. This maps them out in the
new state file, which will allow the removal of select-cri.

The states are split into sections. Parent states are the states that
contain common events used by other states. Shared states are the states
used by more than one journey. And then the journeys that have their own
states (DCMAW, passport, driving licence and f2f). This may not be the
best way to organise them, but we can play with that when we understand
the best way to lay them out.

All the states now contain an attempt-recovery event. This is due to
losing the ATTEMPT_RECOVERY parent state. This state just routed
everything back to the evaluate-gpg45-scores lambda which we are no
longer calling directly. Instead, the event now needs to call back to
the state it's defined on. We've identified this as a piece of tech-debt
that it would be nice to fix up so we no longer need to define it. This
will reduce the size of this file substantially.