PYIC-3013: Pass featureSet to criEvent
The criEvent has a config service. It uses this to determine if CRIs are
enabled or disabled.

The pattern we use for featureSets and the config service is to set
any featureSet being used on the config service itself. As the criEvent
is creating an entirely new config service, it doesn't know about the
the featureSet. So we need to pass it in and set it on it.

An alternative to this would be to use a singleton pattern for the
config service, so that all subsequently instantiated services have that
featureSet info. I have a PR for this.