PYIC-3013: Only return access denied journey if refactor journey
The validate-oauth-callback lambda had logic in it for deciding what
event to send in to the process-journey-step lambda. This was to do with
what version of the multidoc page we should show, if at all.

We've made an assumption that the passport and driving licence CRIs will
always be on. So we can now assume that receiving an `access_denied`
oauth response from DCMAW should always result in showing the multidoc
page. We've also moved the handling of what the "prove your identity
another way" button does to the statemachine.

This all means we can remove the logic and journeys from the lambda. We
preserve the existing responses until we've moved over to the new
journey.