PYIC-3013: Load refactored statemachine fine for all envs
There are no differences between the original journey statemachine files
for dev/build/staging/integration and prod. This means the journeys are
the same without any subtle differences.

The existing refactored statemachine file for dev meets all the feature
tests, and so is functionally the same as the old statemachine files.

This means we can use the same refactored statemachine files for all
envs, rather than maintaining separate copies.

This is good because it will give us confidence that the file ultimately
used in production will have been thoroughly tested as it moves through
the pipelines.

If we want to add new features for development in environments lower
than prod, we can hide them with feature flags and feature sets.

It also adds a couple of tests to ensure that the yaml files are
loadable by the state machine initializer. There is also a test that
walks the resources directory and checks that ANY yaml files are
loadable. This is to hopefully guard against new directory formats being
introduced that would evade the test.

Loading the yaml files in this way gives us some sense that their format
is broadly correct, but it by no means proves that the definition is
"correct" and will behave without error at runtime.