PYIC-3013: Test statemachine classes
We've previously taken an approach of not testing directly the classes
we use to build and operate the statemachine, but instead testing them
implicitly by testing all the steps in the statemachine file. You can
see these tests in ProcessJourneyStepHandlerTest. This file is currently
795 lines long.

Maintaining this testing strategy is not sustainable. There are too many
tests, and they're prone to not being updated. Rather than trying to
test the statemachine file here, we should instead unit test the
"building blocks" of the statemachine system, and then allow the
journeys themselves to be tested with the functional test suite. This
will result in more maintainable tests that are easier to reason about
and hopefully more useful.

Once we've moved over to using the refactored journey statemachine file,
I propose we get rid of the majority of tests in
ProcessJourneyStepHandlerTest.