PYIC-3053: Remove original statemachine files
We've switched over to the refactored journey and no longer need them.

This also removes a lot of tests from ProcessJourneyStepHandlerTest.
These tests were just testing the states and events configured in the
the yaml files. We've got tests that test the domain objects work as
expected, and we have tests that test the journeys themselves. These
tests were cumbersome and hard to maintain.