PYIC-3055: Add issuer to visited cri's and use
To be able to correctly determine that the *last* CRI visited gave us a
VC that was unsuccessful with no CIs we need to use the list of visited
credential issuers. The statuses we generate only contained the CRI
issuer and the visited credenential issuers list only contained the cri
ID. This adds the issuer to the visited credential issuers, so we can
map them together in the eval-gpg45-scores lambda.

Co-authored-by: Laurence <laurence.webb@digital.cabinet-office.gov.uk>