BAU: Conditionally turn off step function logging
This is to address an unexpected issue that is preventing our
CloudFormataion deploying to integration and production.

We need to stop the step functions from logging. Updating their logging
configuration to `OFF` in the template caused an error to be thrown when
deploying step function in the stack:

```
Resource handler returned message: "Model validation failed
(#/LoggingConfiguration/Level: #: only 1 subschema matches out of
2) #/LoggingConfiguration/Level: failed validation constraint for
keyword [enum] (#/LoggingConfiguration/Level)"
```

Changing the logging level to anything else works - ERROR for example -
however we need to stop it completely.

Setting the level to OFF and removing all the other logging properties
(Destination and IncludeExecutionData), also fails to deploy.

Removing the `Logging` property completely will deploy. The docs suggest
that this should then default to logging being off. However after a
deploy with no `Logging` property the actual logging level, checked
through the console, remains unaffected. It stays on whatever level was
previously set.

Given all the above I suggest we make this change. It will check if the
logging level is set to off in the env map, and if it is, remove the
logging sections from the step functions completely. This won't actually
change the logging level due to the issue mentioned above, but we can
manually change it to off while we work out what on earth is happening.
We won't have to worry about it being turned back on by a subsequent
deploy. We've actually already manually turned them off. This will allow
us to make releases and unblock people.