BAU: Return no match if returning user has CIs
If a user has CIs stored in the CI storage service, and they fail the
check on their scores, they should receive a no match response.

The recent changes to the journey engine meant that the gpg45-eval
lambda wasn't being called after a user reset, called from the
check-existing-identity lambda. This meant that a user with CIs could
come back into the system and reach the first CRI, before getting a no
match response due to their existing CIs.

This change slightly updates the logic in the check-existing-identity
lambda to return a fail response if necessary, and maintains existing
behaviour beyond that.