PYIC-3083: Rename refactor journey type
We've now switched to the new refactored journey. We should rename the
new journey back to main.

This copies the new statemachine file with the original name. One this
has been deployed out we can update the parameter through the envs to
use the newly copied file. Once this has happened we can clean up the
`ipv-core-refactor-journey` statemachine file.