PYIC-3164: Remove debug journey
The debug journey is no longer used or works. The new statemachine file
doesn’t contain any debug states, and the old journey didn’t support the
debug journey either. So we should get rid of it.

This removes the build-debug-credential-data lambda. This lambda was
only called by the frontend when displaying the debug page so it can be
dropped.

It also removes the logic from the initialize-ipv-session lambda that
checks if a debug journey has been initiated.