BAU: Add spotless check to pre-commit
It can be very annoying if you forget to run Spotless, only to find out
when your PR checks fail.

This adds a `spotlessCheck` to pre-commit. It will only check, not fix
for you, so you'd still have to manually run `spotlessApply`. This can
be changed if people think it's better for it to make the changes for
you.

The plugin being used is one I wrote and is a repo in my personal
GitHub. If people thing doing this is a good idea we should transfer to
to alphagov, in case I go under a bus.