BAU: Bump sonarqube and use non deprecated gradle task
When running the sonarqube task in GHA we get a deprecation warning that
we should be `sonarqube` task is deprecated and we should use `sonar`
instead. This also bumps the plugin to the latest version.

This is mostly being done in an attempt to fix an issue where the GHA is
failing when running that task:

```
Execution failed for task ':sonarqube'.
> Failed to upload report - An unexpected error occurred. Please try again later.
```

Who knows if this will fix that. Lets find out.