PROBLEM-54: Validate CI codes when retrieving VCs
We recently encountered an issue where a CRI was returning an
unrecognised CI code. We should not be accepting these VCs.

This adds a check in the VC validator to compare all the CI codes in the
VC against our known list. If any CI is not recognised the VC fails
validation.

We're not logging the bad CI code to prevent leaks of sensitive
information.