PROBLEM-54: Check CI codes from storage against known
We've had an issue where a CRI has issued incorrect CI codes. This
resulted in a NPE being thrown in the check existing identity lambda
when trying to check the users CI score.

This change will check the CI exists in the CI map, and if not throw an
exception which is handled.