PROBLEM-58: Handle no status matching VC issuer
We've seen some unhandled errors in the logs. These are generated by a
call to `isVcSuccessful` in the UserIdentityServive. This method will
run through the list of generated VC statuses, and pull out the one that
matches with an CRI issuer value from the users VCs. In theory we should
always have one that matches. Turns out we sometimes don't. When we
don't, the method just calls `orElseThrow` which is unhandled.

This updates it to thrown a specific exception which is then handled.

It's hard to determine how a user gets into this state. I've followed
one affected users journey through the logs. It was a very confused
journey. It looks like they might have had multiple journeys running in
the same session, maybe in different tabs. Also they've hit the back
button at least once as there is an attempt recover in there.

More work might be needed to identify how this happens, but in the mean
time we should be handling the error.