PROBLEM-59: Handle no VC status in build-cri-oauth
We've been throwing unhandled `NoSuchElementException` in the
build-cri-oauth-request lambda. This happens when we try to get the
VC status for an issued VC and can't find it.

This change handles the exception and logs some hopefully useful
information.

I've followed a user journey that encountered this issue in the logs. It
looked completely normal. Hopefully this logging will help to work out
what's going on.