PROBLEM-59: Handle VcStatues being null
We have been throwing NPE's in the build-cri-oauth-request lambda when
a user has at least one credential but their VcStatues from their
session is null.

A user should never be in this state. This catches the NPE and logs some
info that might help us determine what's happening.

For the VcStatues to be null, they will never have been set. We only set
them in two places - the check-existing-identity lambda if the user has
a matched profile (for some reason), and in the eval-gpg45-scores lambda
once a user has a credential.

So, for a user to have a credential their VcStatues to be null should
not be possible. I'm guessing something to do with running two tabs with
the same session cookie.

Either way, we should handle the error and work out what's going on, and
if we can fix it.