PYIC-3195: Remove need for attempt-recovery event
The attempt-recovery event is part of a wider mechanism that detects if
user has tried to do something that has taken them off a "prescribed"
journey. Using the naughty back button for example. Or manually entering
a URL.

In a previous version of the statemachine file, before the refactor, the
attempt-recovery event was just defined on a single parent state and
shared. Since the refactor we've had to define it individually on every
event, and that's bloated the file somewhat.

This small change allows us to handle the attempt-recovery event outside
the usual mechanism and just return the state the user is currently on.
This make the file a lot simpler.