PYIC-3198: Remove `type` from events
We can use the `DEDUCTION` method of deserializing events, rather than
having to use a property in the YAML. This tells the objectmapper to
look at the fields defined to determine what sub-type it should be
deserializing into.

We currently only have one event type, but we could add more if we
need to and this strategy should work.