PYIC-3198: Remove `name` property from states
The states do need to have a name property once deserialized - it's the
value that is stored and fetched from the users session as they move
through the system.

We don't need to duplicate it though. The states are organised in the
YAML under keys, with the key being the state name. We can use this to
populate the name property, rather than having to define it again the
state definition.