BAU: Remove JoureyRequestLambda
This was an abstract class which a few of our lambdas extended. It
implemented the RequestStreamHandler interface and dealt with
serializing/deserializing the IO streams.

Extending this abstract class was causing issues with the powertools
logging. We weren't seeing all the juicy details like `coldStart` and
`functionArn` that we'd hope for. So we're removing it.