BAU: Move journey URIs to separate lib
Sonar was moaning about how we define the URIs we use for journey stuff.
It thinks they're an absolute path to something that might be different
on different on different systems.

We can make it ignore them with suppression annotations, but we only
really want to do this once for each URI. It also makes sense to only
define the URI once for the whole project.

This has resulted in a new lib being created, just to home these URIs
and then pull them in to the different lambdas.

This has stopped short of using the lib in all the test files, where
it's slightly less important.