PYIC-3259: Implement nested journey maps
This allows us to reuse chunks of journey map.

Initially the only real candidate for re-use is the address -> fraud
CRIs journey which is used in a few places.

This changes makes the changes to the state machine initializer, and
adds some supporting classes, to facilitate this. The journey map is
then updated to reuse the new nested journey definition for address and
fraud.