BAU: Fix ContraIndicatorEvidenceDto
The ContraIndicatorEvidenceDto did not match the fields in the CiMit
issued CI and mitigations VC. The VC uses a field in the evidence block
called `contraIndicator`. The dto was looking for `ci`.

The result of this was that when using the CiMit VC we weren't detecting
any CIs that were being returned.