PYIC-3411: Point core dev and build at stubs-prod CiMit stub
This updates the core-back template to have the build and dev
environments use the CiMit stub in the stubs-prod account. It also
updates the naming of some template params to be more consistent.

Devs still have the option to use their own deploy if desired - only the
default has changed.