PYIC-3411: Point core at cimit stub in dev and buld
This has previously been merged and then revered due to the stub not
being ready.

This updates the template to connect the individual dev envs and the
build env with our CiMit stub, rather than SPoTs cimit stub. This should
allow us to start using the VC endpoint in build and so have confidence
in turning it on in higher envs.