SPIKE: Enabled snapstart in dev
This is a spike to see what happens if we use snapstart.
https://docs.aws.amazon.com/lambda/latest/dg/snapstart.html

* It only applies changes in the dev env
* Snapstart only works with x86_64 architecture
* Our step functions weren't invoking our lambdas with a published
  version (via an alias). Instead they were just invoking whatever the
latest version was. You need to specify the version for snapstart to
work, hence the changes to the step function resources in the template.
* There needed to be a change to the TokenRequestValidator. The audience
  value to validate the request against had been baked in to the
snapstart image and was set to the wrong thing. This ensures that the
audience gets set at first runtime.

These changes made the dev experience much...snappier. It might be a
good thing to implement, and if we like it we could maybe even use it in
prod and turn off provisioned concurrency. Do we *need* to run arm64 in
prod though? We'd also need to do a bunch of checking to ensure we
weren't shooting ourselves in the foot like with the audience issue
mentioned above.