SPIKE: Remove calls to parameter store from init
In a few places we were fetching params from SSM in the init of the
lambdas. This was an issue in dev when deploying new stacks from scratch
as it baked in the wrong parameter values to the snapstart image. It
would also mean that the param couldn't be changed until a new version
of the lambda was created so a new snapstart image created.

This also fixes up some stuff around the component ID. We are using a
log helper which was just setting a static value of `core` instead of
the actual component ID from the config.